fix(repo): readd correctly
